oci package passing tests
