+++ helpers_unix_test.go
@@ -7,28 +7,29 @@ import (
 	"fmt"
 
 	"github.com/containerd/containerd/containers"
+	"github.com/containerd/containerd/oci"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 )
 
 const newLine = "\n"
 
-func withExitStatus(es int) SpecOpts {
-	return func(_ context.Context, _ *Client, _ *containers.Container, s *specs.Spec) error {
+func withExitStatus(es int) oci.SpecOpts {
+	return func(_ context.Context, _ oci.Client, _ *containers.Container, s *specs.Spec) error {
 		s.Process.Args = []string{"sh", "-c", fmt.Sprintf("exit %d", es)}
 		return nil
 	}
 }
 
-func withProcessArgs(args ...string) SpecOpts {
-	return WithProcessArgs(args...)
+func withProcessArgs(args ...string) oci.SpecOpts {
+	return oci.WithProcessArgs(args...)
 }
 
-func withCat() SpecOpts {
-	return WithProcessArgs("cat")
+func withCat() oci.SpecOpts {
+	return oci.WithProcessArgs("cat")
 }
 
-func withTrue() SpecOpts {
-	return WithProcessArgs("true")
+func withTrue() oci.SpecOpts {
+	return oci.WithProcessArgs("true")
 }
 
 func withExecExitStatus(s *specs.Process, es int) {
@@ -41,5 +42,5 @@ func withExecArgs(s *specs.Process, args ...string) {
 
 var (
 	withNewSnapshot = WithNewSnapshot
-	withImageConfig = WithImageConfig
+	withImageConfig = oci.WithImageConfig
 )
